Moved the provider implementations out of the 'internal.provider' package and into the 'provider' package. Changed AndroidLocaleInformationProvider so its constructor uses a Context reference instead of one of Configuration.
