+++ sdk/core/azure-core/src/main/java/com/azure/android/core/provider/AndroidPlatformInformationProvider.java
@@ -1,16 +1,14 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.core.internal.provider;
+package com.azure.android.core.provider;
 
 import android.os.Build;
-
-import com.azure.android.core.provider.PlatformInformationProvider;
 
 /**
  * Provider that contains platform information extracted from the {@link Build} class.
  */
-public class AndroidPlatformInformationProvider implements PlatformInformationProvider {
+class AndroidPlatformInformationProvider implements PlatformInformationProvider {
     @Override
     public String getManufacturer() {
         return Build.MANUFACTURER;
