+++ sdk/core/azure-core/src/main/java/com/azure/android/core/provider/ApplicationInformationProvider.java
@@ -5,8 +5,6 @@ package com.azure.android.core.provider;
 
 import android.content.Context;
 
-import com.azure.android.core.internal.provider.AndroidApplicationInformationProvider;
-
 /**
  * Interface for a provider that contains application information.
  */
