+++ sdk/core/azure-core/src/main/java/com/azure/android/core/provider/PlatformInformationProvider.java
@@ -3,8 +3,6 @@
 
 package com.azure.android.core.provider;
 
-import com.azure.android.core.internal.provider.AndroidPlatformInformationProvider;
-
 /**
  * Interface for a provider that contains platform information.
  */
